+++ PSReleaseTools.psm1
@@ -1,6 +1,8 @@
 #load functions
-. $PSScriptRoot\functions\private.ps1
-. $PSScriptRoot\functions\public.ps1
+Get-ChildItem -Path $PSScriptRoot\functions\*.ps1 |
+ForEach-Object {
+    . $_.fullname
+}
 
 #configure TLS settings for GitHub
 [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12